* instead of a single SSHException, now 2 subclasses TransportException and UserAuthException used by pertinent layers * make packet sequence numbers longs, and wrap around to 0 after 2^32 packets * started thoroughly documenting userauth.* * abstract base class for services * made Buffer put*() calls support method chaining * moved classes around (again!), method naming, etc

+++ src/main/java/org/apache/commons/net/ssh/mac/MAC.java
@@ -40,5 +40,5 @@ public interface MAC
     
     void update(byte[] foo, int start, int len);
     
-    void update(int foo);
+    void update(long foo);
 }
